requirements.txt
bcrypt==4.2.0
flask-blueprint==1.3.0
Flask-Cors==4.0.1
Flask-PyMongo==2.3.0
Jinja2==3.1.4
marshmallow==3.21.3
PyJWT==2.8.0
requests==2.32.3
Werkzeug==3.0.3
python-decouple
python-dotenv